requirements.txt
tensorflow
keras
numpy
flask
pillow